requirements.txt
-i https://pypi.org/simple
google-api-python-client==1.7.8
google-cloud-bigquery==1.11.1
confuse==0.5.0
slackclient==1.3.1

